fix: stdlib keyword conflicts and test suite fixes
- Fix do-while continue jump to condition check (not body start)
- Fix Ok/Err case sensitivity: only capitalized forms are Result constructors
- Fix native_bench sum_squares: use long instead of int to avoid overflow
- Fix native_registry parser: handle pub unsafe extern in wrappers.rs
- Fix 19 stdlib .tr files with keyword/identifier conflicts (type, module, is, throw, default)
- Fix 39 bare err() calls -> Err() and 25 bare ok() calls -> Ok()
- Mark native registry test and stdlib_runtest_one as ignored (require LLVM)

+++ trc/tests/native_bench.rs
@@ -79,9 +79,9 @@ fn trc_binary() -> Option<PathBuf> {
 /// (no classes, no ArrayList), and it has a tight loop that benefits
 /// from native codegen.
 const SUM_SQUARES_SOURCE: &str = r#"
-public fn sumSquares(n: int): int {
-    var sum: int = 0;
-    var i: int = 0;
+public fn sumSquares(n: long): long {
+    var sum: long = 0;
+    var i: long = 0;
     while (i < n) {
         sum = sum + i * i;
         i = i + 1;
@@ -90,7 +90,7 @@ public fn sumSquares(n: int): int {
 }
 
 public fn main(): void {
-    let result: int = sumSquares(10000);
+    let result: long = sumSquares(10000);
     io::println(result);
 }
 "#;
